requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
openai==1.12.0
azure-devops==7.1.0b4
langchain==0.1.9
langchain-openai==0.0.5
pydantic==2.6.1
python-multipart==0.0.9
tiktoken==0.6.0